+++ .github/workflows/test.yml
@@ -48,7 +48,7 @@ jobs:
             --ignore=tests/models/test_sevennet.py \
             --ignore=tests/models/test_mattersim.py \
             --ignore=tests/models/test_metatensor.py \
-            --ignore=tests/models/test_torchsim_vs_ase_fire_mace.py \
+            --ignore=tests/test_optimizers_vs_ase.py \
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v5
@@ -69,10 +69,7 @@ jobs:
           - { name: graphpes, test_path: "tests/models/test_graphpes.py" }
           - { name: mace, test_path: "tests/models/test_mace.py" }
           - { name: mace, test_path: "tests/test_elastic.py" }
-          - {
-              name: mace,
-              test_path: "tests/models/test_torchsim_vs_ase_fire_mace.py",
-            }
+          - { name: mace, test_path: "tests/test_optimizers_vs_ase.py" }
           - { name: mattersim, test_path: "tests/models/test_mattersim.py" }
           - { name: metatensor, test_path: "tests/models/test_metatensor.py" }
           - { name: orb, test_path: "tests/models/test_orb.py" }
